[FINAL]Finalized the structure and added a summary report service
