+++ src/app.module.ts
@@ -2,9 +2,12 @@ import { ClassSerializerInterceptor, Module } from '@nestjs/common';
 import { AppController } from './app.controller';
 import { AppService } from './app.service';
 import { APP_INTERCEPTOR } from '@nestjs/core';
+import { SummaryModule } from './summary/summary.module';
+
+import { ReportModule } from './report/report.module';
 
 @Module({
-  imports: [],
+  imports: [SummaryModule, ReportModule],
   controllers: [AppController],
   providers: [
     AppService,
