+++ src/report/report.controller.spec.ts
@@ -1,0 +1,18 @@
+import { Test, TestingModule } from '@nestjs/testing';
+import { ReportController } from './report.controller';
+
+describe('ReportController', () => {
+  let controller: ReportController;
+
+  beforeEach(async () => {
+    const module: TestingModule = await Test.createTestingModule({
+      controllers: [ReportController],
+    }).compile();
+
+    controller = module.get<ReportController>(ReportController);
+  });
+
+  it('should be defined', () => {
+    expect(controller).toBeDefined();
+  });
+});
