+++ src/report/report.module.ts
@@ -1,0 +1,10 @@
+import { Module } from '@nestjs/common';
+import { ReportController } from './report.controller';
+import { ReportService } from './report.service';
+
+@Module({
+  controllers: [ReportController],
+  providers: [ReportService],
+  exports: [ReportService],
+})
+export class ReportModule {}
