+++ src/summary/summary.controller.spec.ts
@@ -1,0 +1,18 @@
+import { Test, TestingModule } from '@nestjs/testing';
+import { SummaryController } from './summary.controller';
+
+describe('SummaryController', () => {
+  let controller: SummaryController;
+
+  beforeEach(async () => {
+    const module: TestingModule = await Test.createTestingModule({
+      controllers: [SummaryController],
+    }).compile();
+
+    controller = module.get<SummaryController>(SummaryController);
+  });
+
+  it('should be defined', () => {
+    expect(controller).toBeDefined();
+  });
+});
