+++ src/summary/summary.controller.ts
@@ -1,0 +1,12 @@
+import { Controller, Get } from '@nestjs/common';
+import { SummaryService } from './summary.service';
+
+@Controller('summary')
+export class SummaryController {
+  constructor(private readonly summaryService: SummaryService) {}
+
+  @Get('')
+  getSummary() {
+    return this.summaryService.calculateSummary();
+  }
+}
