+++ src/summary/summary.service.ts
@@ -1,0 +1,24 @@
+import { Injectable } from '@nestjs/common';
+import { ReportType } from 'src/data';
+import { ReportService } from 'src/report/report.service';
+
+@Injectable()
+export class SummaryService {
+  constructor(private readonly reportService: ReportService) {}
+
+  calculateSummary() {
+    const totalExpense = this.reportService
+      .getAllReports(ReportType.EXPENSE)
+      .reduce((sum, report) => sum + report.amount, 0);
+
+    const totalIncome = this.reportService
+      .getAllReports(ReportType.INCOME)
+      .reduce((sum, report) => sum + report.amount, 0);
+
+    return {
+      totalIncome,
+      totalExpense,
+      netIncome: totalIncome - totalExpense,
+    };
+  }
+}
